requirements.txt
aioice==0.6.18
aiortc==0.9.27
av==7.0.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
crc32c==2.0
cryptography==2.8
idna==2.9
netifaces==0.10.9
picamera==1.13
pkg-resources==0.0.0
pycparser==2.20
pyee==7.0.1
pyFirmata2==2.0.1
pylibsrtp==0.6.6
pyserial==3.4
python-dotenv==0.12.0
python-engineio==3.12.1
python-socketio==4.4.0
requests==2.23.0
six==1.14.0
urllib3==1.25.8
